Move explorer components from src/components to src/explorer_common
