+++ client/src/explorer_common/explorer-components.scss
